Move `RemoteFeedImageDataLoader` to production
